Reworked the ReminderVariable class to use long instead of int and all other classes that depended on it. Also go Room to work Hazza!

+++ app/src/main/java/com/example/ma18ea/ReminderVariables.kt
@@ -2,7 +2,7 @@ package com.example.ma18ea
 
 class ReminderVariables(
         var title:String = "newTitle",
-        var doneTime:Int = 0,
-        var totalTime:Int = 0,
+        var doneTime:Long = 0,
+        var totalTime:Long = 0,
         var days:ArrayList<String> = ArrayList(7),
         var description:String = "Description") {}
